MOB-1759: Implement BookmarksExchange and refactor code
