+++ Client/Frontend/Library/Bookmarks/BookmarksPanel.swift
@@ -591,35 +591,8 @@ extension BookmarksPanel {
     }
 
     func exportBookmarksActionHandler(_ action: UIAlertAction) {
-        let activityIndicator = UIActivityIndicatorView(style: .medium)
-        activityIndicator.color = UIColor.theme.ecosia.primaryBrand
-        activityIndicator.startAnimating()
-        
-        let toast = SimpleToast().showAlertWithText(
-            "Exporting Bookmarks…",
-            image: .view(activityIndicator),
-            bottomContainer: view,
-            dismissAfter: nil,
-            bottomInset: view.layoutMargins.bottom
-        )
-        
-        toast.onShown = { [weak self] in
-            Task {
-                guard let self = self else { return }
-                let serializer = BookmarkSerializer()
-                
-                let items = try await self.viewModel.getBookmarksForExport()
-                
-                let htmlExport = serializer.serializeBookmarks(items)
-                
-                let exportedBooksmarksUrl = FileManager.default.temporaryDirectory.appendingPathComponent("Bookmarks.html")
-                try htmlExport.data(using: .utf8)?.write(to: exportedBooksmarksUrl)
-
-                let activityViewController = UIActivityViewController(activityItems: [exportedBooksmarksUrl], applicationActivities: nil)
-                self.present(activityViewController, animated: true) {
-                    toast.dismiss()
-                }
-            }
+        Task { @MainActor in
+            try await viewModel.bookmarkExportSelected(in: self)
         }
     }
